Move things into library, finally
