+++ packages/vfx-composer/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./Particles"
+export * from "./units"
+export * from "./useParticles"
